feat: organize cli commands by type
Commands related to references are now organized as subcommands of "ref" command

Commands related to HMMs are now organized as subcommands of "hmm" command

update documentation
